refactor : Generator 패키지 변경
