+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import nextstep.ladder.generator.NonConsecutiveFlagGenerator;
 
 public class Ladder {
