+++ src/main/java/nextstep/ladder/domain/Lines.java
@@ -2,6 +2,7 @@ package nextstep.ladder.domain;
 
 import java.util.List;
 import java.util.Objects;
+import nextstep.ladder.generator.NonConsecutiveFlagGenerator;
 
 public class Lines {
