+++ src/main/java/nextstep/ladder/generator/NonConsecutiveFlagGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.generator;
 
 import java.util.List;
 import java.util.stream.IntStream;
